requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
boto3==1.26.99
botocore==1.29.99
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
dateparser==1.1.8
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.7
django-allauth==0.53.1
django-cors-headers==3.14.0
django-heroku==0.3.1
django-storages==1.13.2
djangorestframework==3.14.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
jet-bridge-base==1.7.10
jmespath==1.0.1
oauthlib==3.2.2
Pillow==9.4.0
promise==2.3
prompt-toolkit==2.0.9
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
Rx==1.6.3
s3transfer==0.6.0
six==1.16.0
social-auth-core==4.4.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
wcwidth==0.2.6
whitenoise==6.4.0
django-cloudinary-storage
django-admin-soft-dashboard